修改了退票方法的一个问题，本来退款金额被错误地指定为整数

+++ test/test/test/Program.cs
@@ -985,7 +985,7 @@ namespace test
 
                 // 5. 执行退票
                 decimal refundFee;
-                int refundAmount;
+                decimal refundAmount;
                 bool success = _bookingService.RefundTicket(
                     selectedOrder.OrderID,
                     //DateTime.Now,
